Move PyPI release testing into own workflow
