+++ .github/workflows/pypi.yml
@@ -1,0 +1,48 @@
+name: PyPI release testing of baked projects
+
+on:
+  # This test takes a very long time to complete, but will only be affected by very
+  # specific changes or upstream changes. We therefore restrict this test to manual
+  # triggers and cron jobs.
+  workflow_dispatch:
+  schedule:
+  - cron: "0 5 * * 1"
+
+jobs:
+  test-pypi-release:
+    name: Testing PyPI release
+    runs-on: ubuntu-18.04
+
+    steps:
+    - name: Checking out the cookie cutter repository
+      uses: actions/checkout@v2
+
+    - name: Set up Python ${{ matrix.python-version }}
+      uses: actions/setup-python@v2
+      with:
+        python-version: 3.8
+
+    - name: Install test requirements
+      run: |
+        python -m pip install -r requirements-dev.txt
+
+    - name: Set up git identity
+      run: |
+        git config --global user.email "ssc@citestuser.com"
+        git config --global user.name "SSC CI Test User"
+
+    - name: Set up SSH Agent to deploy to test repositories
+      uses: webfactory/ssh-agent@v0.4.1
+      with:
+        ssh-private-key: |
+          ${{ secrets.GHA_TEST_PRIVATE_KEY }}
+
+    - name: Test deploying the baked project to Github + Gitlab.com
+      run: |
+        python -m pytest -m deploy
+
+    - name: Run tests for upstream integrations
+      env:
+        GH_API_ACCESS_TOKEN: ${{ secrets.GH_API_ACCESS_TOKEN }}
+      run: |
+        python -m pytest -m pypi
