+++ pytest.ini
@@ -4,3 +4,4 @@ markers =
     deploy: marks tests that require a deploy key setup to be run
     integrations: marks tests that require previous run of deploy tests and might need API keys set up
     local: marks tests that only perform local operations
+    pypi: marks expensive tests that build wheels in CI
